Fix responsivity of primary-header
Change the way the padding of the page is made.

The side padding was keeping showing up in the viewport until the specification of "@media(max-width)", with the content being ugly shrunk, and the image was being cropped. I needed the side padding of the content of the page to vanish with the decreasing of the width of the screen even before the "@media" alteration.

I replaced the using of padding directly in "root" for a "min-width" in "primary-header", altering the width of "root" to 100vw, and creating an overlay to make the background of "primary-header". I also modified the value of "@media(max-width)", and made both sides of "billboard-content" have the same width.