Remove page bottom scroll bar (responsivity)
Stop using viewport-based metrics, because it was making a bottom scroll bar to show up.

I replaced "vw" for "%", and limitated each part of the "root" elements, instead of doing it to all "root", which permitted me to make the backgrounds that ought to fill entire rows of the page todo so.